scx_bpfland: override wakee's L3 domain with waker's in WAKE_SYNC
In the CPU idle selection process for the WAKE_SYNC case, if the waker
is not an exiting task, override the wakee's L3 domain with that of the
waker. This allows the wakee to migrate to the waker's domain,
optimizing producer->consumer pipeline performance.

Moreover, return more meaningful error codes from pick_idle_cpu() (this
is only to improve code readability).

Signed-off-by: Andrea Righi <andrea.righi@linux.dev>