scx_fair: prevent DSQ starvation
Always wake up the CPU of the first task from any DSQ whenever a new
task is added. This ensures that tasks in the per-CPU DSQs are never
starved, as it keeps their CPU active and ready to process tasks: this
ensures that tasks in the per-CPU DSQs are not starved when CPUs are
going idle and not being awakened.

Additionally, make sure to provide opportunities to consume tasks from
both the per-CPU DSQ and the shared DSQ in ops.dispatch() to avoid
starvation: this ensures that tasks in the shared DSQ are not starved by
the per-CPU DSQs.

Lastly, automatically migrate tasks queued in per-CPU DSQs to the shared
DSQ if they cannot be processed by the CPU associated with the per-CPU
DSQ they are queued in: this allows to fix stalls due to tasks being
dispatched to the wrong per-CPU DSQ due to race conditions with their
cpumask or prev_cpu.

Signed-off-by: Andrea Righi <andrea.righi@linux.dev>